function_call/requirements.txt
flask==3.0.3
flask-cors==6.0.1
flask-socketio==5.5.1
python-socketio==5.13.0
requests==2.31.0
python-dotenv==1.0.0
numpy>=1.26.0,<2.0.0
openai==1.3.0